Excercise2
