+++ SK14 Automation Java/src/Lecture2/homework1.java
@@ -1,0 +1,12 @@
+package Lecture2;
+public class homework1 {
+    // Store your name in 3 different variables and print the whole name
+    public static void main(String [] args) {
+        String firstName = "Tsveta";
+        String middleName = "Valentinova";
+        String lastName = "Vasileva";
+
+        System.out.println(firstName + " " + middleName + " " + lastName);
+    }
+}
+
