+++ SK14 Automation Java/src/Lecture2/homework2.java
@@ -1,12 +1,13 @@
 package Lecture2;
-public class homework2 {
-    // Store your name in 3 different variables and print the whole name
-    public static void main(String [] args) {
-        String firstName = "Tsveta";
-        String middleName = "Valentinova";
-        String lastName = "Vasileva";
 
-        System.out.println(firstName + " " + middleName + " " + lastName);
+public class homework2 {
+    // Print the perimeter of a triangle by given sides. Use formula to calculate it.
+    public static void main(String[] args) {
+        int a = 3;
+        int b = 3;
+        int c = 3;
+
+        int perimeter = a + b + c;
+        System.out.println("Perimeter is: " + perimeter);
     }
 }
-
