+++ SK14 Automation Java/src/Lecture2/homework4.java
@@ -1,15 +1,18 @@
 package Lecture2;
 
 public class homework4 {
-    // Print the area of a triangle by given sides. Use formula to calculate it.
-    public static void main(String[] args) {
 
-        double sideA = 8.0;
-        double sideB = 10.0;
-        double sideC = 10.0;
-        double semiPerimeter = (sideA + sideB + sideC) / 2;
-
-        double area = Math.sqrt(semiPerimeter * (semiPerimeter - sideA) * (semiPerimeter - sideB) * (semiPerimeter - sideC));
-        System.out.println("The area of the triangle is: " + area);
+    public static void main (String[] args) {
     }
-}
+
+    {
+        int treeHeight = 5;
+        for (int i = 0; i < treeHeight; i++) {
+            for (int j = 0; j < treeHeight - i - 1; j++) {
+                System.out.print(" ");
+            }
+            for (int k = 0; k < 2 * i + 1; k++) {
+                System.out.print("*");
+            }
+            System.out.println();
+        }}}
