+++ SK14 Automation Java/src/Lecture2/homework6.java
@@ -1,24 +1,18 @@
 package Lecture2;
 
-import java.util.Scanner;
-
 public class homework6 {
+    public static void main(String[] args) {
+        double a = 1.0;
+        double b = -3.0;
+        double c = 2.0;
 
-    public static class InputExample {
-        public static void main (String[] args) {
-            String firstName = "Tsveta";
-            String middleName = "Valentinova";
-            String lastName = "Vasileva";
+        double discriminant = b * b - 4 * a * c;
 
-            Scanner scanner = new Scanner(System.in);
+        double root1 = (-b + Math.sqrt(discriminant)) / (2 * a);
+        double root2 = (-b - Math.sqrt(discriminant)) / (2 * a);
 
-            System.out.print(firstName + " T " + middleName + " V " + lastName + " V ");
-            int number = scanner.nextInt();
+        System.out.println("Root 1: " + root1);
+        System.out.println("Root 2: " + root2);
+    }
+}
 
-            int result = number * 2;
-
-            System.out.println("The result is: " + result);
-
-            scanner.close();
-        }
-    }}
